move strategist utils into e2e utils
